+++ src/voku/SimplePhpParser/Model/PHPTrait.php
@@ -31,11 +31,11 @@ final class PHPTrait extends BasePHPClass
 
         // Extract PHP 8.0+ attributes
         if (!empty($node->attrGroups)) {
-            $this->attributes = Utils::extractAttributesFromAstNode($node->attrGroups);
+            $this->attributes = Utils::extractAttributesFromAstNode($node->attrGroups, $this->parserContainer);
         }
 
         $traitExists = false;
-        if (self::canAutoloadFromPhpNode($node)) {
+        if ($this->reflectionEnrichmentEnabled() && self::canAutoloadFromPhpNode($node)) {
             try {
                 if (\trait_exists($this->name, true)) {
                     $traitExists = true;